Remove all remaining `l10n` and `LocalizationHelper` references
