+++ src/devtools/client/inspector/markup/utils/l10n.js
@@ -4,6 +4,4 @@
 
 "use strict";
 
-const { LocalizationHelper } = require("devtools/shared/l10n");
-
 module.exports = {};
