+++ src/devtools/client/webconsole/components/Output/MessageRepeat.js
@@ -7,7 +7,6 @@
 const PropTypes = require("prop-types");
 const dom = require("react-dom-factories");
 const { PluralForm } = require("devtools/shared/plural-form");
-const { l10n } = require("devtools/client/webconsole/utils/messages");
 const messageRepeatsTooltip = "#1 repeat;#1 repeats";
 
 MessageRepeat.displayName = "MessageRepeat";
