+++ src/devtools/client/webconsole/utils/messages.js
@@ -4,7 +4,6 @@
 
 "use strict";
 
-const l10n = require("devtools/client/webconsole/utils/l10n");
 /*
 const {
   getUrlDetails,
@@ -464,6 +463,5 @@ module.exports = {
   isGroupType,
   isError,
   isBrowserInternalMessage,
-  l10n,
   prepareMessage,
 };
