+++ src/devtools/packages/devtools-reps/reps/grip-array.js
@@ -218,7 +218,6 @@ function arrayIterator(props, grip, max) {
 }
 
 function getEmptySlotsElement(number) {
-  // TODO: Use l10N - See https://github.com/firefox-devtools/reps/issues/141
   return `<${number} empty slot${number > 1 ? "s" : ""}>`;
 }
