--- src/devtools/shared/l10n.js
@@ -1,2 +1,0 @@
-import { LocalizationHelper } from "shims/l10n";
-export { LocalizationHelper };
